// src/app/shared/interfaces/otros.interface.ts
// Generated by https://quicktype.io

export interface Gasolineras {
  nombre:        string;
  direccion:     string;
  poblacion:     string;
  pais:          string;
  coordenadas:   Coordenadas;
  gMaps:         string;
  todoElDia:     boolean;
}

// Generated by https://quicktype.io

export interface Tiendas {
  nombre:       string;
  tipo:         string;
  direccion:    string;
  poblacion:    string;
  coordenadas:  Coordenadas;
  gMaps:        string;
}

export interface Coordenadas {
  latitud:   number;
  longitud:  number;
}